requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
itsdangerous
jinja2
python-multipart


# agent creation
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langgraph
pydantic
langchain-tavily

markdown